requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-cleanup==8.0.0
django-smart-selects==1.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
